Force a date (if it is only a year) into a string instead of a number
When generating the manifest, the metadata is transformed as XML and
then encoded as JSON. Because dates (as in a single year) looks like
a number and we use JSON_NUMERIC_CHECK with json_encode (because the
width and height of the image should be numbers), they will be encoded
as a number. To prevent this from happening, we append a non-breaking-
space to the date.